# examples/next15-app/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the OpenAPI specification for your project."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "x-generator": {
    "apiDir": "./src/app/api",
    "schemaDir": "./src",
    "outputFile": "swagger.json",
    "onlyOpenApiRoutes": false
  },
  "paths": {
    "/todos": {
      "get": {
        "operationId": "get-todos",
        "summary": "",
        "description": "",
        "tags": [
          "Todos"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-todos",
        "summary": "",
        "description": "",
        "tags": [
          "Todos"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/todos/{id}": {
      "get": {
        "operationId": "get-todos-[id]",
        "summary": "",
        "description": "",
        "tags": [
          "Todos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "operationId": "put-todos-[id]",
        "summary": "",
        "description": "",
        "tags": [
          "Todos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete-todos-[id]",
        "summary": "",
        "description": "",
        "tags": [
          "Todos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  }
}